fix: redesign desktop layout, responsive container width, and rich executive home dashboard

+++ server/api/health.get.ts
@@ -5,8 +5,8 @@ export default defineEventHandler(async () => {
   const db = useDatabase()
   const result = await db.execute(sql`
     SELECT now() AS waktu,
-           (SELECT count(*) FROM employees)   AS pegawai,
-           (SELECT count(*) FROM leave_types) AS jenis_izin
+           (SELECT count(*)::int FROM employees)   AS pegawai,
+           (SELECT count(*)::int FROM leave_types) AS jenis_izin
   `)
   return { data: { database: 'ok', ...(result[0] as Record<string, unknown>) } }
 })
